% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObtainPooledFits.R
\name{.ObtainPooledFits}
\alias{.ObtainPooledFits}
\title{Pool the fits for imputed dataset regressions}
\usage{
.ObtainPooledFits(
  dataset,
  variable_list,
  sdq = c("h_sdq_external", "h_sdq_internal"),
  conf_list
)
}
\arguments{
\item{dataset}{A data.frame containing all stacked imputed datasets}

\item{variable_list}{A data.frame containing a list of exposures for wich regressions will be run, first column must contain exposure names}

\item{sdq}{A character vector defining SDQ score}

\item{conf_list}{A character vector specifying confounders to be added in the regression model}
}
\value{
A data.frame containing an untidy regression fits for all exposures
}
\description{
Pool the fits for imputed dataset regressions
}
