% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunRegression.R
\name{.RunRegression}
\alias{.RunRegression}
\title{Run adjusted regression}
\usage{
.RunRegression(input_data, variable_list, sdq, subcohort = NULL, conf_list)
}
\arguments{
\item{input_data}{"mids" object (imputed dataset)}

\item{variable_list}{A data.frame containing a list of exposures for which regressions will be run, first column must contain exposure names}

\item{sdq}{A character vector indicating in the outcome SDQ score is external or internal}

\item{subcohort}{A character vector specifying stratification to be made (default = NULL)}

\item{conf_list}{A character vector specifying confounders to be added in the regression model}
}
\value{
A data.frame with regression results for all exposures in a non-tidy format
}
\description{
Run adjusted regression
}
