% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyExwasAllOutcomes.R
\name{.TidyExwasAllOutcomes}
\alias{.TidyExwasAllOutcomes}
\title{Tidy the outcomes of regressions run for all exposures in a mids object}
\usage{
.TidyExwasAllOutcomes(
  input_data,
  variable_list,
  exwas_result,
  subcohort,
  save = FALSE,
  path = NULL,
  file_name = NULL
)
}
\arguments{
\item{input_data}{mids object (imputed dataset)}

\item{variable_list}{A data.frame containing a list of exposures for wich regressions will be run, first column must contain exposure names}

\item{exwas_result}{Regression output}

\item{subcohort}{A one element character vector indicating a stratification factor}

\item{save}{A character vector specifying if the files are going to be saved (default = TRUE, save files)}

\item{path}{A string defining path for saving}

\item{file_name}{File name (to which .csv or _plot.csv will be added)}
}
\value{
A data.frame with tidy regression result, ready for plotting
}
\description{
Tidy the outcomes of regressions run for all exposures in a mids object
}
