% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyExwasSingleResult.R
\name{.TidyExwasSingleResult}
\alias{.TidyExwasSingleResult}
\title{Generate tidy ExWAS result from a regression on a single exposure}
\usage{
.TidyExwasSingleResult(x)
}
\arguments{
\item{x}{Pooled result of a regression run on a mids object}
}
\value{
Statistics (regression coefficient, 95% CI, p-value...) for the modeled variable of interest
}
\description{
Generate tidy ExWAS result from a regression on a single exposure
}
