% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionSex.R
\name{InteractionSex}
\alias{InteractionSex}
\title{helper function to calculate interaction between sex and exposures identified in the main ExWAS}
\usage{
InteractionSex(input_data, sdq, conf_list, strat_var_list)
}
\arguments{
\item{input_data}{A mids object (imputed dataset)}

\item{sdq}{A string characterizing which SDQ score is to be used "h_sdq_external" or "h_sdq_internal"}

\item{conf_list}{A character vector listing the adjustment factors}

\item{strat_var_list}{A character vector listing the names of variables for which interaction with sex will be checked}
}
\value{
A dataframe with p value for interaction between sex and exposure of interest in an ExWAS adjusted for confounding factors
}
\description{
helper function to calculate interaction between sex and exposures identified in the main ExWAS
}
