% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMidsResults.R
\name{PlotMidsResults}
\alias{PlotMidsResults}
\title{Plots regressions estimates obtained in ExWAS}
\usage{
PlotMidsResults(data_for_plotting, title, plot_name, path, ...)
}
\arguments{
\item{data_for_plotting}{Results of the ExWAS analysis containing: Exposure, Estimate, conf_low and conf_high}

\item{title}{Title of the plot}

\item{plot_name}{Name that the plot will be saved under (of the format: "plot.jpg")}

\item{path}{Path where to save the plot}

\item{...}{other arguments passed to ggplot argument}
}
\value{
A plot of the regression coefficients IRR with 95% confidence intervals
}
\description{
Plots regressions estimates obtained in ExWAS
}
