% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStratifiedResults.R
\name{PlotStratifiedResults}
\alias{PlotStratifiedResults}
\title{Plots regressions estimates obtained in stratified ExWAS}
\usage{
PlotStratifiedResults(
  data_for_plotting,
  colors,
  labels,
  ylim,
  legend_pos,
  path,
  file_name
)
}
\arguments{
\item{data_for_plotting}{Results of the ExWAS analysis containing: Exposure, Estimate, conf_low and conf_high}

\item{colors}{Provide colors to be used for each stratum}

\item{labels}{Provide names to be used for each stratum}

\item{ylim}{A numeric vector defining y axis limits (c(min, max))}

\item{legend_pos}{A numeric vector defining legend position (c(x, y))}

\item{path}{A string defining the path for saving the plot}

\item{file_name}{A string defining the file name for saving the plot}
}
\value{
A plot of stratified regression
}
\description{
Plots regressions estimates obtained in stratified ExWAS
}
