% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExwasCompleteCase.R
\name{RunExwasCompleteCase}
\alias{RunExwasCompleteCase}
\title{Run an ExWAS on a dataset that does not need pooling afterwards (complete case dataset)}
\usage{
RunExwasCompleteCase(
  input_data,
  variable_list,
  sdq = c("h_sdq_external", "h_sdq_internal"),
  conf_list,
  subcohort = NULL,
  save = FALSE,
  path = NULL,
  file_name = NULL
)
}
\arguments{
\item{input_data}{complete "mids" long format (stacked) imputed dataset or any dataset with no pooling required (complete cases for example)}

\item{variable_list}{A data.frame containing a list of exposures for wich regressions will be run, first column must contain exposure names}

\item{sdq}{A character vector indicating in the outcome SDQ score is external or internal}

\item{conf_list}{A character vector defining a lsit of confounders used in the model}

\item{subcohort}{A character vector specifying stratification to be made (one cohort left)}

\item{save}{A character vector specifying if the files are going to be saved (default = TRUE, save files)}

\item{path}{Path for saving (default = NULL)}

\item{file_name}{File name (to which .csv or _plot.csv will be added, default = NULL)}
}
\value{
Regression coefficients (IRR) with '95%' confidence intervals for complete dataset
}
\description{
Run an ExWAS on a dataset that does not need pooling afterwards (complete case dataset)
}
