% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExwasImputed.R
\name{RunExwasImputed}
\alias{RunExwasImputed}
\title{Run an ExWAS on an inputed `mids` object}
\usage{
RunExwasImputed(
  input_data,
  variable_list,
  sdq,
  conf_list,
  subcohort = NULL,
  save = FALSE,
  path = NULL,
  file_name = NULL
)
}
\arguments{
\item{input_data}{"mids" object (imputed dataset)}

\item{variable_list}{A data.frame containing a list of exposures for which regressions will be run, first column must contain exposure names}

\item{sdq}{A character vector indicating in the outcome SDQ score is external or internal}

\item{conf_list}{A character vector specifying confounders to be added in the regression model}

\item{subcohort}{A character vector specifying stratification to be made (default = NULL)}

\item{save}{A character vector specifying if the files are going to be saved (default = FALSE, do not save files)}

\item{path}{Path for saving (default = NULL)}

\item{file_name}{File name (to which .csv or _plot.csv will be added, default = NULL)}
}
\value{
Regression coefficients (IRR) with 95% confidence intervals for imputed dataset
}
\description{
Run an ExWAS on an inputed `mids` object
}
