% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExwasSdqConfounders.R
\name{RunExwasSdqConfounders}
\alias{RunExwasSdqConfounders}
\title{Runs ExWAS for one SDQ score at a time and a confounder}
\usage{
RunExwasSdqConfounders(
  input_data,
  sdq,
  vars_to_regress = NULL,
  conf_list,
  variable_names,
  path,
  file_name
)
}
\arguments{
\item{input_data}{"mids" object (imputed dataset)}

\item{sdq}{A character vector indicating in the outcome SDQ score is external or internal}

\item{vars_to_regress}{character vector containing names of variables that will be included in the multiple regression}

\item{conf_list}{A character vector containing names of adjusting factors}

\item{variable_names}{A data.frame containing a list of exposures for wich regressions will be run, first column must contain exposure names}

\item{path}{Path for saving}

\item{file_name}{Name under which the result will be saved (*.csv)}
}
\value{
Regression coefficients (IRR) with 95% confidence intervals for one confounder vs. SDQ score at a time
}
\description{
Runs ExWAS for one SDQ score at a time and a confounder
}
