% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSignificant.R
\name{SelectSignificant}
\alias{SelectSignificant}
\title{Selects variables that were significant on desired p value level}
\usage{
SelectSignificant(data, pvalue)
}
\arguments{
\item{data}{A dataframe which is a result of ExWAS analysis}

\item{pvalue}{A scalar defining p value threshold of significance}
}
\value{
A character vector containing names of significant exposures
}
\description{
Selects variables that were significant on desired p value level
}
