% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratifyGainWeight.R
\name{StratifyGainWeight}
\alias{StratifyGainWeight}
\title{Calculate regression estimate (ExWAS) for each category of data stratified on weight gain}
\usage{
StratifyGainWeight(data, strat_var_list, variable_list)
}
\arguments{
\item{data}{A data frame of complete dataset}

\item{strat_var_list}{A character vector containing the names of variables to stratify on}

\item{variable_list}{A data.frame containing a list of exposures for which regressions will be run, first column must contain exposure names}
}
\value{
A dataframe with regression estimates with p values for each exposure of interest
}
\description{
Calculate regression estimate (ExWAS) for each category of data stratified on weight gain
}
